Create deposit_request representer and twin
- Move twins and representers under StellarBase namespace
