--- app/twins/application_twin.rb
@@ -1,3 +1,0 @@
-class ApplicationTwin < Disposable::Twin
-  feature Expose
-end
