+++ app/twins/stellar_base/application_twin.rb
@@ -1,0 +1,5 @@
+module StellarBase
+  class ApplicationTwin < Disposable::Twin
+    feature Expose
+  end
+end
